+++ client/src/lib/navigation/constants/routes.ts
@@ -8,13 +8,7 @@ import {
 } from '../../../../../shared/src/types/Session';
 import {CompletedSessionEvent} from '../../../../../shared/src/types/Event';
 
-export type ProfileStackProps = {
-  Profile: undefined;
-  EarlyAccessInfo?: {showBack: boolean};
-};
-
 export type TabNavigatorProps = {
-  ProfileStack: NavigatorScreenParams<ProfileStackProps>;
   Sessions: undefined;
   Journey: undefined;
 };
@@ -45,7 +39,7 @@ export type OverlayStackProps = {
   AboutEditorialOverlay: undefined;
   CommunityEditorialOverlay: undefined;
   AboutOverlay: undefined;
-  CommunityOverlay: undefined;
+  EarlyAccessInfoOverlay?: {showBack: boolean};
 };
 
 export type ModalStackProps = {
