--- content/src/ui/Screen.Profile.json
@@ -1,37 +1,0 @@
-{
-  "en": {
-    "settings": "Settings",
-    "community": "Community",
-    "noDisplayName": "Add your name",
-    "about": "29k Sessions",
-    "actions": {
-      "signIn": "Sign in",
-      "profileSettings": "Profile settings",
-      "notifications": "Automatic reminders",
-      "deleteData": "Delete my data",
-      "developer": "Developer",
-      "language": "Language"
-    }
-  },
-  "pt": {
-    "settings": "Configurações",
-    "community": "Comunidade",
-    "noDisplayName": "Adiciona o teu nome",
-    "about": "Sessões 29k FJN",
-    "actions": {
-      "signIn": "Entrar",
-      "profileSettings": "Configurações do perfil",
-      "deleteData": "Remover os meus dados",
-      "developer": "Programador",
-      "language": "Idioma",
-      "notifications": "Notificações"
-    }
-  },
-  "sv": {
-    "actions": {
-      "signIn": "Logga in",
-      "language": "Språk"
-    }
-  },
-  "es": {}
-}
